+++ src/main/java/edu/yuriiknowsjava/xpinjection/conferences/repositories/ConferenceRepository.java
@@ -1,6 +1,5 @@
 package edu.yuriiknowsjava.xpinjection.conferences.repositories;
 
-import java.math.BigInteger;
 import java.util.Optional;
 
 import edu.yuriiknowsjava.xpinjection.conferences.entities.Conference;
@@ -9,8 +8,8 @@ import org.springframework.data.repository.ListPagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ConferenceRepository extends JpaRepository<Conference, BigInteger>,
-        ListPagingAndSortingRepository<Conference, BigInteger> {
+public interface ConferenceRepository extends JpaRepository<Conference, Long>,
+        ListPagingAndSortingRepository<Conference, Long> {
 
     Optional<Conference> findByName(String name);
