+++ src/main/java/edu/yuriiknowsjava/xpinjection/conferences/repositories/ThemeRepository.java
@@ -1,6 +1,5 @@
 package edu.yuriiknowsjava.xpinjection.conferences.repositories;
 
-import java.math.BigInteger;
 import java.util.Set;
 
 import edu.yuriiknowsjava.xpinjection.conferences.entities.Theme;
@@ -8,7 +7,7 @@ import org.springframework.data.repository.ListPagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ThemeRepository extends ListPagingAndSortingRepository<Theme, BigInteger> {
+public interface ThemeRepository extends ListPagingAndSortingRepository<Theme, Long> {
 
-    Set<Theme> findByIdInOrTagIn(Set<BigInteger> ids, Set<String> names);
+    Set<Theme> findByIdInOrTagIn(Set<Long> ids, Set<String> names);
 }
